backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
pydantic==2.9.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
openai==1.50.0  
anthropic==0.34.2
pypdf==3.17.4  # Better than PyPDF2
python-docx==1.1.2
pgvector==0.3.3
pdfplumber==0.10.3  # Additional PDF extraction
python-dotenv==1.0.0
tiktoken==0.5.2  # Token counting for OpenAI
aiofiles==23.2.1  # Async file operations